split and updated organization for the event bus system
